move world view widget to generic folder
