+++ src/software/gui/generic_widgets/world_view/BUILD
@@ -1,0 +1,24 @@
+load("@bazel_rules_qt//:qt.bzl", "qt_cc_library")
+
+package(default_visibility = ["//software/gui/visualizer:__subpackages__"])
+
+qt_cc_library(
+    name = "world_view",
+    src = "world_view.cpp",
+    hdr = "world_view.h",
+    deps = [
+        "//software/gui/generic_widgets/world_view:zoomable_qgraphics_view",
+        "@qt//:qt_core",
+        "@qt//:qt_widgets",
+    ],
+)
+
+qt_cc_library(
+    name = "zoomable_qgraphics_view",
+    src = "zoomable_qgraphics_view.cpp",
+    hdr = "zoomable_qgraphics_view.h",
+    deps = [
+        "@qt//:qt_gui",
+        "@qt//:qt_widgets",
+    ],
+)
