+++ src/software/gui/generic_widgets/world_view/zoomable_qgraphics_view.h
