+++ src/software/gui/visualizer/widgets/BUILD
@@ -9,11 +9,11 @@ qt_cc_library(
     deps = [
         ":ai_control",
         ":parameters",
-        ":world_view",
         "//software/ai/hl/stp:play_info",
         "//software/backend:robot_status",
         "//software/geom",
         "//software/gui/generic_widgets/robot_status",
+        "//software/gui/generic_widgets/world_view",
         "//software/gui/visualizer/drawing:draw_functions",
         "//software/gui/visualizer/geom:geometry_conversion",
         "//software/gui/visualizer/ui:main_widget",
@@ -40,23 +40,12 @@ qt_cc_library(
 )
 
 qt_cc_library(
-    name = "world_view",
-    src = "world_view.cpp",
-    hdr = "world_view.h",
-    deps = [
-        ":zoomable_qgraphics_view",
-        "@qt//:qt_core",
-        "@qt//:qt_widgets",
-    ],
-)
-
-qt_cc_library(
     name = "parameters",
     src = "parameters.cpp",
     hdr = "parameters.h",
     deps = [
-        ":zoomable_qgraphics_view",
         "//software/gui/generic_widgets/robot_status:robot_status_table",
+        "//software/gui/generic_widgets/world_view:zoomable_qgraphics_view",
         "//software/gui/visualizer/ui:main_widget",
         "//software/logger",
         "//software/parameter",
@@ -69,22 +58,12 @@ qt_cc_library(
     src = "ai_control.cpp",
     hdr = "ai_control.h",
     deps = [
-        ":zoomable_qgraphics_view",
         "//software/ai/hl/stp/play",
         "//software/gui/generic_widgets/robot_status:robot_status_table",
+        "//software/gui/generic_widgets/world_view:zoomable_qgraphics_view",
         "//software/gui/visualizer/ui:main_widget",
         "//software/parameter:dynamic_parameters",
         "//software/util/design_patterns:generic_factory",
-        "@qt//:qt_widgets",
-    ],
-)
-
-qt_cc_library(
-    name = "zoomable_qgraphics_view",
-    src = "zoomable_qgraphics_view.cpp",
-    hdr = "zoomable_qgraphics_view.h",
-    deps = [
-        "@qt//:qt_gui",
         "@qt//:qt_widgets",
     ],
 )
